Utils creation for validation utilities and refactors
