+++ onboardProject/pages/LoginPage.tsx
@@ -4,42 +4,7 @@ import {H1} from '../atomic/atm/atm.typo/typo.style'
 import { TextField } from '../atomic/atm/atm.input/input.component'
 import { View, Alert } from 'react-native';
 import PageContainer from '../atomic/atm/atm.pagecontainer/pagecontainer.component';
-
-interface authPack {
-    email: string;
-    password: string;
-};
-
-const validateLogin = (pack:authPack) => {
-    const minimumSize = (pack.password.length >= 7);
-
-    if (!validateEmail(pack.email)){
-        Alert.alert("Email inválido " + pack.email)
-    }
-    else if (!validatePassword(pack.password)){
-        Alert.alert("Pelo menos 1 número e 1 caractere")
-    }
-    else if (!minimumSize){
-        Alert.alert("Senha muito curta (7 min.)")
-    }
-    else{
-        Alert.alert("Sucesso");
-    }
-}
-
-const validateEmail = (email: string) => {
-    const regexValidator = /.+[@].+\.com/;
-    const valid = regexValidator.test(email);
-    
-    return valid;
-}
-
-const validatePassword = (password: string) => {
-    const regexValidator = /((.*[A-Z].*)|(.*[a-z].*))(.*[0-9].*)/;
-    const valid = regexValidator.test(password);
-    
-    return valid;
-}
+import { validateLogin } from 'utils/validationUtils';
 
 export const LoginPage = () => 
 {
